+++ confluence/embedded-confluence-public/docs/00-intro.tsx
@@ -1,0 +1,21 @@
+import { md } from '@atlaskit/docs';
+
+export default md`
+  ## Introduction
+
+  This package provides components (ViewPage and EditPage) to any parent product which needs to embed confluence experience within the parent products.
+
+  ### What is Embedded Pages?
+
+  Embedded Pages is a lightweight Confluence experience that allows users of parent product such as JSM, JSW, Trello etc. to edit/view long form page content, where the experience stays completely within the parent product. In addition, we plan to insert Embedded pages into external products such as GitLab, MS Teams, Zoom, Slack, etc.
+
+  ### Goals of Embedded Pages:
+
+  _1. Deliver value and promote “better together” product experience_ - allow parent product users to easily create long form content and collaborate/share knowledge within the context of that parent product, and thereby deriving sufficient value from this windowed experience.
+
+  _2. Drive awareness of Confluence_ - showcase the power/potential of Confluence as your go-to technology of choice for your content creation and collaboration needs, and entice users to upgrade and use the stand alone Confluence app.
+
+  ### What’s in it for the customer?
+
+  We have frequently heard from customers that the experience among Atlassian products are somewhat broken/disjointed which leads to user friction and increases fear of using our products. With Embedded Pages, parent product users can directly edit/view long form content within the same product context that they are working in. This eliminates the constant context-switching between the parent product and Confluence.
+`;