+++ confluence/embedded-confluence-public/docs/02-components-of-Embedded-Pages.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { md } from '@atlaskit/docs';
+
+import Base from './base-document/base';
+
+export default md`
+    ${(
+      <Base
+        content={require('!!raw-loader!./markdown-files/components-of-embedded-pages.md')}
+      />
+    )}
+`;