+++ confluence/embedded-confluence-public/docs/base-document/base.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import ReactMarkdown from 'react-markdown';
+
+export default function Base(props: any) {
+  return <ReactMarkdown children={props.content} />;
+}