+++ navigation/atlassian-navigation/docs/ert/skeleton-create-button.tsx
@@ -1,0 +1,3 @@
+import { SkeletonCreateButtonProps } from '../../src/components/SkeletonCreateButton/types';
+
+export default (props: SkeletonCreateButtonProps) => null;