+++ navigation/atlassian-navigation/docs/ert/skeleton-icon-button.tsx
@@ -1,0 +1,3 @@
+import { SkeletonIconButtonProps } from '../../src/components/SkeletonIconButton/types';
+
+export default (props: SkeletonIconButtonProps) => null;