+++ navigation/atlassian-navigation/docs/ert/skeleton-primary-button.tsx
@@ -1,0 +1,3 @@
+import { SkeletonPrimaryButtonProps } from '../../src/components/SkeletonPrimaryButton/types';
+
+export default (props: SkeletonPrimaryButtonProps) => null;