+++ navigation/atlassian-navigation/docs/ert/skeleton-switcher-button.tsx
@@ -1,0 +1,3 @@
+import { SkeletonSwitcherButtonProps } from '../../src/skeleton-switcher-button';
+
+export default (props: SkeletonSwitcherButtonProps) => null;