+++ navigation/atlassian-navigation/src/skeleton-help-button.tsx
@@ -5,7 +5,15 @@ import QuestionCircleIcon from '@atlaskit/icon/glyph/question-circle';
 import { SkeletonIconButton } from './components/SkeletonIconButton';
 import { useTheme } from './theme';
 
-export const SkeletonHelpButton = ({ label = '' }: { label: string }) => {
+export type SkeletonHelpButtonProps = {
+  /**
+   *  Describes the specific role of this navigation component for users viewing the page with a screen
+   *  reader. Differentiates from other navigation buttons on a page.
+   */
+  label: string;
+};
+
+export const SkeletonHelpButton = ({ label = '' }: SkeletonHelpButtonProps) => {
   const {
     mode: { navigation },
   } = useTheme();