+++ navigation/atlassian-navigation/src/skeleton-settings-button.tsx
@@ -4,7 +4,17 @@ import SettingsIcon from '@atlaskit/icon/glyph/settings';
 
 import { SkeletonIconButton } from './components/SkeletonIconButton';
 
-export const SkeletonSettingsButton = ({ label = '' }: { label: string }) => (
+export type SkeletonSettingsButtonProps = {
+  /**
+   *  Describes the specific role of this navigation component for users viewing the page with a screen
+   *  reader. Differentiates from other navigation buttons on a page.
+   */
+  label: string;
+};
+
+export const SkeletonSettingsButton = ({
+  label = '',
+}: SkeletonSettingsButtonProps) => (
   <SkeletonIconButton>
     <SettingsIcon label={label} />
   </SkeletonIconButton>