+++ navigation/atlassian-navigation/src/skeleton-switcher-button.tsx
@@ -4,7 +4,17 @@ import AppSwitcherIcon from '@atlaskit/icon/glyph/app-switcher';
 
 import { SkeletonIconButton } from './components/SkeletonIconButton';
 
-export const SkeletonSwitcherButton = ({ label = '' }: { label: string }) => (
+export type SkeletonSwitcherButtonProps = {
+  /**
+   *  Describes the specific role of this navigation component for users viewing the page with a screen
+   *  reader. Differentiates from other navigation buttons on a page.
+   */
+  label: string;
+};
+
+export const SkeletonSwitcherButton = ({
+  label = '',
+}: SkeletonSwitcherButtonProps) => (
   <SkeletonIconButton>
     <AppSwitcherIcon label={label} />
   </SkeletonIconButton>