+++ confluence/legacy-mobile-macros/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/legacy-mobile-macros
 
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 2.0.8
 
 ### Patch Changes