+++ design-system/checkbox/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/checkbox
 
+## 12.3.13
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 12.3.12
 
 ### Patch Changes