+++ design-system/comment/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/comment
 
+## 10.5.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.5.1
 
 ### Patch Changes