+++ design-system/drawer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/drawer
 
+## 7.1.10
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 7.1.9
 
 ### Patch Changes