+++ design-system/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/dropdown-menu
 
+## 11.1.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 11.1.4
 
 ### Patch Changes