+++ design-system/focus-ring/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/focus-ring
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 1.0.4
 
 ### Patch Changes