+++ design-system/lozenge/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/lozenge
 
+## 11.1.11
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 11.1.10
 
 ### Patch Changes