+++ design-system/page-layout/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/page-layout
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 1.2.2
 
 ### Patch Changes