+++ design-system/progress-bar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/progress-bar
 
+## 0.5.6
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.5.5
 
 ### Patch Changes