+++ design-system/range/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/range
 
+## 6.0.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 6.0.4
 
 ### Patch Changes