+++ design-system/section-message/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/section-message
 
+## 6.1.12
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 6.1.11
 
 ### Patch Changes