+++ design-system/tag/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/tag
 
+## 11.3.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 11.3.2
 
 ### Patch Changes