+++ design-system/textfield/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/textfield
 
+## 5.1.11
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 5.1.10
 
 ### Patch Changes