+++ design-system/toggle/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/toggle
 
+## 12.4.8
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 12.4.7
 
 ### Patch Changes