+++ dev-portal/rating/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/rating
 
+## 0.2.15
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.2.14
 
 ### Patch Changes