+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @atlaskit/editor-mobile-bridge
 
+## 34.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
+## 34.0.2
+
+### Patch Changes
+
+- Updated dependencies
+
+## 34.0.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 34.0.0
 
 ### Minor Changes