+++ elements/user-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/user-picker
 
+## 9.0.7
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 9.0.6
 
 ### Patch Changes