+++ help/help/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/help
 
+## 7.1.8
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 7.1.7
 
 ### Patch Changes